# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Bluba Anticipa API
  version: 0.1.0-mvp
  summary: Contrato unificado Frontend móvil ↔ Backend ↔ motor predictivo para NeuroHack 2026.
  description: 'API del MVP de Bluba Anticipa. Define los contratos externos consumidos por la app móvil y un contrato interno
    opcional para invocar el motor predictivo. El sistema es una herramienta de apoyo preventivo; los scores del MVP son índices
    demostrativos y no probabilidades clínicas validadas.


    Principios del contrato: control de acceso por rol, riesgo y confianza separados, explicabilidad, datos faltantes explícitos,
    confirmación humana de observaciones estructuradas y recomendaciones provenientes de fuentes controladas.'
servers:
- url: http://localhost:8080
  description: Desarrollo local
- url: https://api.example.bluba.cl
  description: Placeholder de despliegue del MVP
...
security:
- bearerAuth: []
paths:
  /v1/auth/session:
    post:
      tags:
      - Auth
      operationId: createDemoSession
      summary: Crear sesión demostrativa por rol
      description: Endpoint del MVP. No pretende modelar autenticación productiva completa.
      security: []
      x-traceability:
        user_stories:
        - HU-TR-01
        requirements:
        - RF-TR-01
        - RF-TR-02
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DemoSessionRequest'
      responses:
        '200':
          description: Sesión creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/me/context:
    get:
      tags:
      - Context
      operationId: getMyContext
      summary: Obtener rol, permisos y recursos autorizados
      x-traceability:
        user_stories:
        - HU-TR-01
        - HU-TR-02
        - HU-COL-04
        requirements:
        - RF-TR-03
        - RF-TR-04
        - RF-COL-04
      responses:
        '200':
          description: Contexto del usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserContext'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/children:
    get:
      tags:
      - Children
      operationId: listAuthorizedChildren
      summary: Listar niños/pacientes autorizados
      x-traceability:
        user_stories:
        - HU-TR-02
        requirements:
        - RF-TR-04
        - RF-TR-05
      responses:
        '200':
          description: Lista filtrada por autorización.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChildSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/children/{childId}/preventive-status:
    get:
      tags:
      - Children
      operationId: getPreventiveStatus
      summary: Obtener estado preventivo compuesto adaptado al rol
      parameters:
      - $ref: '#/components/parameters/ChildId'
      x-traceability:
        user_stories:
        - HU-TR-03
        - HU-FAM-07
        - HU-FAM-08
        - HU-FAM-10
        - HU-EDU-04
        requirements:
        - RF-TR-06
        - RF-TR-07
        - RF-PRED-15
      responses:
        '200':
          description: Estado preventivo vigente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreventiveStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/children/{childId}/risk-predictions/current:
    get:
      tags:
      - Predictions
      operationId: getCurrentRiskPrediction
      summary: Obtener predicción vigente de riesgo y confianza
      description: En INSUFFICIENT_DATA, risk.score y risk.level son null. Los niveles son calculados por Backend/Data-ML;
        el Frontend no replica thresholds.
      parameters:
      - $ref: '#/components/parameters/ChildId'
      x-traceability:
        user_stories:
        - HU-TR-04
        - HU-TR-05
        - HU-TR-06
        - HU-TR-07
        - HU-ESP-08
        - HU-ESP-09
        - HU-ESP-10
        requirements:
        - RF-PRED-07
        - RF-PRED-08
        - RF-PRED-10
        - RF-PRED-12
        - RF-PRED-13
        - RF-PRED-14
      responses:
        '200':
          description: Predicción actual o estado de insuficiencia.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskPrediction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/children/{childId}/daily-records:
    post:
      tags:
      - Capture
      operationId: createDailyRecord
      summary: Crear check-in/registro diario confirmado
      parameters:
      - $ref: '#/components/parameters/ChildId'
      - $ref: '#/components/parameters/IdempotencyKey'
      x-traceability:
        user_stories:
        - HU-FAM-01
        requirements:
        - RF-CAP-01
        - RF-FAM-01
        - RF-PRED-15
        - RNF-REL-04
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DailyRecordCreateRequest'
      responses:
        '201':
          description: Registro persistido; el recálculo puede continuar de forma asincrónica.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyRecordResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /v1/children/{childId}/dysregulation-events:
    post:
      tags:
      - Events
      operationId: createDysregulationEvent
      summary: Registrar una escalada o desregulación con payload mínimo
      parameters:
      - $ref: '#/components/parameters/ChildId'
      - $ref: '#/components/parameters/IdempotencyKey'
      x-traceability:
        user_stories:
        - HU-EDU-06
        - HU-EDU-07
        - HU-EDU-08
        requirements:
        - RF-EDU-06
        - RF-EDU-07
        - RF-EDU-08
        - RNF-OFF-01
        - RNF-REL-04
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DysregulationEventCreateRequest'
      responses:
        '201':
          description: Evento aceptado/persistido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DysregulationEventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de sesión. En el MVP puede ser un token demostrativo; la autorización siempre se aplica en Backend.
    serviceBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token servicio-a-servicio para el endpoint interno del motor predictivo.
  parameters:
    ChildId:
      name: childId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
      example: CH-002
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Recomendado para escrituras originadas desde el cliente móvil/offline. Repetir la misma clave no debe crear
        duplicados.
      schema:
        type: string
        minLength: 8
        maxLength: 128
  schemas:
    Role:
      type: string
      enum:
      - FAMILY
      - EDUCATOR
      - PROFESSIONAL
    UsageContext:
      type: string
      enum:
      - HOME
      - SCHOOL
      - PROFESSIONAL
    RecordSource:
      type: string
      enum:
      - FAMILY
      - SCHOOL
      - PROFESSIONAL
    PersistenceStatus:
      type: string
      enum:
      - PERSISTED
      - PENDING_SYNC
      - SYNCED
    RiskLevel:
      type:
      - string
      - 'null'
      enum:
      - LOW
      - MEDIUM
      - HIGH
      - null
    ConfidenceLevel:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
    PredictionStatus:
      type: string
      enum:
      - OK
      - LOW_CONFIDENCE
      - INSUFFICIENT_DATA
      - ERROR
    BaselineStatus:
      type: string
      enum:
      - BUILDING
      - PROVISIONAL
      - SUFFICIENT
    FactorDirection:
      type: string
      enum:
      - INCREASES_RISK
      - DECREASES_RISK
      - NEUTRAL
    FactorType:
      type: string
      enum:
      - DEVIATION
      - ACCUMULATION
      - CONTEXT
      - HISTORY
      - DATA_QUALITY
    MissingDataState:
      type: string
      enum:
      - MISSING
      - STALE
    WarningSeverity:
      type: string
      enum:
      - INFO
      - WARNING
      - ERROR
...
    RecommendationSource:
      type: string
      enum:
      - PROFESSIONAL_APPROVED
      - HISTORICAL_SUCCESS
      - VALIDATED_CATALOG
...
    ObservationProvenance:
      type: string
      enum:
      - HUMAN_STRUCTURED
      - AI_EXTRACTED_HUMAN_CONFIRMED
    DysregulationEventType:
      type: string
      enum:
      - ESCALATION
      - DYSREGULATION
    ProblemDetails:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          example: FORBIDDEN_CHILD_ACCESS
        message:
          type: string
          example: No tiene acceso al niño solicitado.
        trace_id:
          type:
          - string
          - 'null'
        field_errors:
          type: array
          items:
            type: object
            required:
            - field
            - message
            properties:
              field:
                type: string
              message:
                type: string
          default: []
      additionalProperties: false
    DemoSessionRequest:
      type: object
      required:
      - role
      properties:
        role:
          $ref: '#/components/schemas/Role'
        demo_user_key:
          type:
          - string
          - 'null'
          description: Selector opcional de identidad demo.
      additionalProperties: false
    SessionResponse:
      type: object
      required:
      - access_token
      - token_type
      - user_id
      - role
      properties:
        access_token:
          type: string
        token_type:
          type: string
          const: Bearer
        expires_in_seconds:
          type: integer
          minimum: 60
          default: 3600
        user_id:
          type: string
        role:
          $ref: '#/components/schemas/Role'
      additionalProperties: false
    ChildSummary:
      type: object
      required:
      - id
      - display_name
      properties:
        id:
          type: string
        display_name:
          type: string
        relationship_contexts:
          type: array
          items:
            $ref: '#/components/schemas/UsageContext'
      additionalProperties: false
    UserContext:
      type: object
      required:
      - user_id
      - role
      - children
      properties:
        user_id:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        children:
          type: array
          items:
            $ref: '#/components/schemas/ChildSummary'
        classrooms:
          type: array
          items:
            type: object
            required:
            - id
            - name
            properties:
              id:
                type: string
              name:
                type: string
            additionalProperties: false
          default: []
        permissions:
          type: array
          items:
            type: string
          default: []
      additionalProperties: false
    SleepQuality:
      type:
      - string
      - 'null'
      enum:
      - reparador
      - interrumpido
      - dificultad_conciliacion
      - desconocido
      - null
    WakeState:
      type:
      - string
      - 'null'
      enum:
      - tranquilo_alegre
      - irritable_llorando
      - cansado_sueno
      - desconocido
      - null
    RegulationLevel:
      type:
      - string
      - 'null'
      enum:
      - excelente
      - estable_con_apoyo
      - desregulacion_frecuente
      - desconocido
      - null
    AlertLevel:
      type:
      - string
      - 'null'
      enum:
      - bajo
      - optimo
      - alto
      - desconocido
      - null
    CanonicalDailyFeatures:
      type: object
      properties:
        sleep_quality:
          $ref: '#/components/schemas/SleepQuality'
        sleep_hours:
          type:
          - number
          - 'null'
          minimum: 0
          maximum: 24
        wake_state:
          $ref: '#/components/schemas/WakeState'
        regulation_level:
          $ref: '#/components/schemas/RegulationLevel'
        alert_level:
          $ref: '#/components/schemas/AlertLevel'
        routine_change:
          type:
          - boolean
          - 'null'
        gastrointestinal_status:
          type:
          - string
          - 'null'
          maxLength: 64
        observed_behavior:
          type: array
          items:
            type: string
          default: []
        exceptional_event:
          type:
          - boolean
          - 'null'
        sensory_profile:
          type: array
          items:
            type: string
          default: []
      additionalProperties: false
    DailyRecordCreateRequest:
      type: object
      required:
      - recorded_at
      - source
      - features
      properties:
        recorded_at:
          type: string
          format: date-time
        source:
          $ref: '#/components/schemas/RecordSource'
        context:
          $ref: '#/components/schemas/UsageContext'
        features:
          $ref: '#/components/schemas/CanonicalDailyFeatures'
        notes:
          type:
          - string
          - 'null'
          maxLength: 2000
        provenance:
          $ref: '#/components/schemas/ObservationProvenance'
      additionalProperties: false
    DailyRecordResponse:
      type: object
      required:
      - record_id
      - child_id
      - recorded_at
      - source
      - persistence_status
      - risk_recalculation_requested
      properties:
        record_id:
          type: string
        child_id:
          type: string
        recorded_at:
          type: string
          format: date-time
        source:
          $ref: '#/components/schemas/RecordSource'
        persistence_status:
          $ref: '#/components/schemas/PersistenceStatus'
        risk_recalculation_requested:
          type: boolean
        risk_recalculation_request_id:
          type:
          - string
          - 'null'
      additionalProperties: false
...
    RiskValue:
      type:
      - object
      - 'null'
      required:
      - score
      - level
      properties:
        score:
          type:
          - number
          - 'null'
          minimum: 0
          maximum: 1
          description: Índice interno del motor. No implica probabilidad clínica calibrada.
        level:
          $ref: '#/components/schemas/RiskLevel'
      additionalProperties: false
    ConfidenceValue:
      type: object
      required:
      - score
      - level
      properties:
        score:
          type: number
          minimum: 0
          maximum: 1
        level:
          $ref: '#/components/schemas/ConfidenceLevel'
      additionalProperties: false
    TopFactor:
      type: object
      required:
      - code
      - label
      properties:
        code:
          type: string
          example: SLEEP_ALTERED_3D
        label:
          type: string
          example: Sueño alterado durante 3 días
        direction:
          $ref: '#/components/schemas/FactorDirection'
        window:
          type:
          - string
          - 'null'
          example: 72h
        type:
          $ref: '#/components/schemas/FactorType'
        contribution:
          type:
          - number
          - 'null'
          description: Magnitud explicativa opcional propia del motor; no se interpreta como causalidad.
      additionalProperties: false
    MissingDataItem:
      type: object
      required:
      - field
      - state
      properties:
        field:
          type: string
          example: sleep_quality
        state:
          $ref: '#/components/schemas/MissingDataState'
        priority:
          type: integer
          minimum: 1
          maximum: 5
          default: 3
        reason:
          type:
          - string
          - 'null'
      additionalProperties: false
    DataQuality:
      type: object
      required:
      - completeness
      - history_days
      - sources
      - missing_fields
      properties:
        completeness:
          type: number
          minimum: 0
          maximum: 1
        critical_present:
          type:
          - integer
          - 'null'
          minimum: 0
        critical_total:
          type:
          - integer
          - 'null'
          minimum: 0
        hours_since_last_record:
          type:
          - number
          - 'null'
          minimum: 0
        history_days:
          type: integer
          minimum: 0
        sources:
          type: array
          items:
            $ref: '#/components/schemas/RecordSource'
        missing_fields:
          type: array
          items:
            type: string
        missing_critical_data:
          type: array
          items:
            $ref: '#/components/schemas/MissingDataItem'
          default: []
        contains_synthetic_data:
          type: boolean
          default: false
      additionalProperties: false
    PredictionWarning:
      type: object
      required:
      - code
      - severity
      - message
      properties:
        code:
          type: string
        severity:
          $ref: '#/components/schemas/WarningSeverity'
        message:
          type: string
      additionalProperties: false
    RiskPrediction:
      type: object
      required:
      - prediction_id
      - child_id
      - prediction_at
      - window_end_at
      - horizon_hours
      - model_version
      - feature_schema_version
      - status
      - risk
      - confidence
      - top_factors
      - data_quality
      - warnings
      properties:
        prediction_id:
          type: string
        child_id:
          type: string
        prediction_at:
          type: string
          format: date-time
        window_end_at:
          type: string
          format: date-time
        horizon_hours:
          type: integer
          const: 24
        model_version:
          type: string
          example: hybrid-mvp-v1
        feature_schema_version:
          type: string
          example: features-mvp-v1
        status:
          $ref: '#/components/schemas/PredictionStatus'
        risk:
          $ref: '#/components/schemas/RiskValue'
        confidence:
          $ref: '#/components/schemas/ConfidenceValue'
        top_factors:
          type: array
          items:
            $ref: '#/components/schemas/TopFactor'
        data_quality:
          $ref: '#/components/schemas/DataQuality'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/PredictionWarning'
        required_fields:
          type: array
          items:
            type: string
          default: []
      additionalProperties: false
...
    BaselineMetric:
      type: object
      required:
      - key
      - kind
      properties:
        key:
          type: string
        kind:
          type: string
          enum:
          - NUMERIC
          - CATEGORICAL
        median:
          type:
          - number
          - 'null'
        robust_dispersion:
          type:
          - number
          - 'null'
        mode:
          type:
          - string
          - 'null'
        distribution:
          type: object
          additionalProperties:
            type: number
          default: {}
        valid_observations:
          type: integer
          minimum: 0
      additionalProperties: false
    BaselineSnapshot:
      type: object
      required:
      - child_id
      - status
      - valid_days
      - reference_window_days
      - cutoff_at
      - metrics
      properties:
        child_id:
          type: string
        status:
          $ref: '#/components/schemas/BaselineStatus'
        valid_days:
          type: integer
          minimum: 0
        reference_window_days:
          type: integer
          const: 14
        cutoff_at:
          type: string
          format: date-time
        metrics:
          type: array
          items:
            $ref: '#/components/schemas/BaselineMetric'
      additionalProperties: false
...
    Recommendation:
      type: object
      required:
      - strategy_id
      - name
      - priority
      - source
      - contexts
      - reason
      properties:
        strategy_id:
          type: string
        name:
          type: string
        priority:
          type: integer
          minimum: 1
        source:
          $ref: '#/components/schemas/RecommendationSource'
        contexts:
          type: array
          items:
            type: string
            enum:
            - HOME
            - SCHOOL
            - BOTH
        reason:
          type: string
        instructions:
          type:
          - string
          - 'null'
        historical_evidence:
          type:
          - object
          - 'null'
          properties:
            times_used:
              type: integer
              minimum: 0
            successful_or_partial:
              type: integer
              minimum: 0
          additionalProperties: false
      additionalProperties: false
...
    DysregulationEventCreateRequest:
      type: object
      required:
      - occurred_at
      - event_type
      - context
      properties:
        occurred_at:
          type: string
          format: date-time
        event_type:
          $ref: '#/components/schemas/DysregulationEventType'
        context:
          $ref: '#/components/schemas/UsageContext'
        notes:
          type:
          - string
          - 'null'
          maxLength: 2000
        trigger_tags:
          type: array
          items:
            type: string
          default: []
        strategy_id:
          type:
          - string
          - 'null'
      additionalProperties: false
    DysregulationEventResponse:
      type: object
      required:
      - event_id
      - child_id
      - occurred_at
      - event_type
      - context
      - persistence_status
      properties:
        event_id:
          type: string
        child_id:
          type: string
        occurred_at:
          type: string
          format: date-time
        event_type:
          $ref: '#/components/schemas/DysregulationEventType'
        context:
          $ref: '#/components/schemas/UsageContext'
        persistence_status:
          $ref: '#/components/schemas/PersistenceStatus'
        risk_recalculation_requested:
          type: boolean
          default: true
      additionalProperties: false
...
    PreventiveStatus:
      type: object
      required:
      - child_id
      - generated_at
      - prediction
      - baseline
      - recommendations
      - disclaimer
      properties:
        child_id:
          type: string
        generated_at:
          type: string
          format: date-time
        prediction:
          $ref: '#/components/schemas/RiskPrediction'
        baseline:
          $ref: '#/components/schemas/BaselineSnapshot'
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/Recommendation'
          maxItems: 5
        disclaimer:
          type: string
          default: Estimación preventiva de apoyo; no corresponde a un diagnóstico.
      additionalProperties: false
...
  responses:
    BadRequest:
      description: Solicitud inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: No autenticado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: No autorizado para el recurso solicitado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: Conflicto de estado o idempotencia.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ServerError:
      description: Error interno.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'